model/requirements.txt
kaggle
python-dotenv
numpy
nltk
scipy